requirements.txt
numpy~=2.2.4
torch~=2.6.0
wandb~=0.19.8
tqdm~=4.67.1
trl~=0.16.0
transformers~=4.50.3
PyYAML~=6.0.2
pillow~=11.1.0
peft